SGREP_TMPDIR. sgrep.exe.

+++ sakura_grep/getopt.cpp
@@ -39,7 +39,7 @@ static void postpone(int argc, WCHAR* const argv[], int index) {
 static int postpone_noopt(int argc, WCHAR* const argv[], int index) {
     int i = index;
     for(; i < argc; i++) {
-        if(*(argv[i]) == L'-') {
+        if (*(argv[i]) == L'-') {
             postpone(argc, argv, index);
             return 1;
         }
@@ -48,33 +48,33 @@ static int postpone_noopt(int argc, WCHAR* const argv[], int index) {
 }
 static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const struct option* longopts, int* longindex)
 {
-    while(1) {
+    while (1) {
         int c;
         const WCHAR* optptr = 0;
-        if(optind >= argc - postpone_count) {
+        if (optind >= argc - postpone_count) {
             c = 0;
             optarg = 0;
             break;
         }
         c = *(argv[optind] + nextchar);
-        if(c == L'\0') {
+        if (c == L'\0') {
             nextchar = 0;
             ++optind;
             continue;
         }
-        if(nextchar == 0) {
-            if(optstring[0] != L'+' && optstring[0] != L'-') {
-                while(c != L'-') {
+        if (nextchar == 0) {
+            if (optstring[0] != L'+' && optstring[0] != L'-') {
+                while (c != L'-') {
                     /* postpone non-opt parameter */
-                    if(!postpone_noopt(argc, argv, optind)) {
+                    if (!postpone_noopt(argc, argv, optind)) {
                         break; /* all args are non-opt param */
                     }
                     ++postpone_count;
                     c = *argv[optind];
                 }
             }
-            if(c != L'-') {
-                if(optstring[0] == L'-') {
+            if (c != L'-') {
+                if (optstring[0] == L'-') {
                     optarg = argv[optind];
                     nextchar = 0;
                     ++optind;
@@ -82,22 +82,22 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
                 }
                 break;
             } else {
-                if(wcscmp(argv[optind], L"--") == 0) {
+                if (wcscmp(argv[optind], L"--") == 0) {
                     optind++;
                     break;
                 }
                 ++nextchar;
-                if(longopts != 0 && *(argv[optind] + 1) == L'-') {
+                if (longopts != 0 && *(argv[optind] + 1) == L'-') {
                     WCHAR const* spec_long = argv[optind] + 2;
                     WCHAR const* pos_eq = wcschr(spec_long, L'=');
-                    int spec_len = (pos_eq == NULL ? wcslen(spec_long) : pos_eq - spec_long);
+                    size_t spec_len = (pos_eq == NULL ? wcslen(spec_long) : pos_eq - spec_long);
                     int index_search = 0;
                     int index_found = -1;
                     const struct option* optdef = 0;
-                    while(longopts->name != 0) {
-                        if(wcsncmp(spec_long, longopts->name, spec_len) == 0) {
-                            if(optdef != 0) {
-                                if(opterr) {
+                    while (longopts->name != 0) {
+                        if (wcsncmp(spec_long, longopts->name, spec_len) == 0) {
+                            if (optdef != 0) {
+                                if (opterr) {
                                     fwprintf(stderr, L"ambiguous option: %s\n", spec_long);
                                 }
                                 return L'?';
@@ -108,8 +108,8 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
                         longopts++;
                         index_search++;
                     }
-                    if(optdef == 0) {
-                        if(opterr) {
+                    if (optdef == 0) {
+                        if (opterr) {
                             fwprintf(stderr, L"no such a option: %s\n", spec_long);
                         }
                         return L'?';
@@ -117,15 +117,15 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
                     switch(optdef->has_arg) {
                         case no_argument:
                             optarg = 0;
-                            if(pos_eq != 0) {
-                                if(opterr) {
+                            if (pos_eq != 0) {
+                                if (opterr) {
                                     fwprintf(stderr, L"no argument for %s\n", optdef->name);
                                 }
                                 return L'?';
                             }
                             break;
                         case required_argument:
-                            if(pos_eq == NULL) {
+                            if (pos_eq == NULL) {
                                 ++optind;
                                 optarg = argv[optind];
                             } else {
@@ -135,10 +135,10 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
                     }
                     ++optind;
                     nextchar = 0;
-                    if(longindex != 0) {
+                    if (longindex != 0) {
                         *longindex = index_found;
                     }
-                    if(optdef->flag != 0) {
+                    if (optdef->flag != 0) {
                         *optdef->flag = optdef->val;
                         return 0;
                     }
@@ -148,42 +148,40 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
             }
         }
         optptr = wcschr(optstring, c);
-        if(optptr == NULL) {
+        if (optptr == NULL) {
             optopt = c;
-            if(opterr) {
+            if (opterr) {
                 fwprintf(stderr,
-                        L"%s: invalid option -- %c\n",
-                        argv[0], c);
+                         L"%s: invalid option -- %c\n",
+                         argv[0], c);
             }
             ++nextchar;
             return L'?';
         }
-        if(*(optptr+1) != L':') {
+        if (*(optptr+1) != L':') {
             nextchar++;
-            if(*(argv[optind] + nextchar) == L'\0') {
+            if (*(argv[optind] + nextchar) == L'\0') {
                 ++optind;
                 nextchar = 0;
             }
             optarg = 0;
         } else {
             nextchar++;
-            if(*(argv[optind] + nextchar) != L'\0') {
+            if (*(argv[optind] + nextchar) != L'\0') {
                 optarg = argv[optind] + nextchar;
             } else {
                 ++optind;
-                if(optind < argc - postpone_count) {
+                if (optind < argc - postpone_count) {
                     optarg = argv[optind];
                 } else {
                     optopt = c;
-                    if(opterr) {
+                    if (opterr) {
                         fwprintf(stderr,
                             L"%s: option requires an argument -- %c\n",
                             argv[0], c);
                     }
-                    if(optstring[0] == L':' ||
-                        (optstring[0] == L'-' || optstring[0] == L'+') &&
-                        optstring[1] == L':')
-                    {
+					if (optstring[0] == L':' ||
+                       (optstring[0] == L'-' || optstring[0] == L'+') && optstring[1] == L':') {
                         c = L':';
                     } else {
                         c = L'?';
@@ -199,7 +197,7 @@ static int _getopt_(int argc, WCHAR* const argv[], const WCHAR* optstring, const
     /* end of option analysis */
 
     /* fix the order of non-opt params to original */
-    while((argc - optind - postpone_count) > 0) {
+    while ((argc - optind - postpone_count) > 0) {
         postpone(argc, argv, optind);
         ++postpone_count;
     }
